August Challenge
In HackerChess, there is a special piece, known as the VROOK. A VROOK is like a normal Rook, except that it can move only along the column. A VROOK can't capture/kill another VROOK, and it also can't jump over another VROOK.

HackerChess is played as follows. The board is an NxN grid. It is played between two players. Each column of the board contains exactly 2 VROOKs, one of player-1 and the other of player-2. Each player has N VROOKS, arranged on the board. In a turn, a player can move any of his VROOKs, (i.e) if there are any possible VROOKs belonging to the player that has a position to which it can be moved. The player to make the last move wins.

Given an initial configuration of the board. If both player-1 and player-2 play optimally. Who will win the game, if player-2 gets to move first?